chore: renamed AnimeDialogResultCard.tsx to AnimeSearchResultCard.tsx
